Split tests accoding to what they test
