+++ vaadin-dev-server/src/test/java/com/vaadin/base/devserver/stats/ProjectHelpersTest.java
@@ -1,0 +1,61 @@
+package com.vaadin.base.devserver.stats;
+
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNotSame;
+
+import java.io.File;
+import java.io.IOException;
+
+import com.vaadin.flow.testutil.TestUtils;
+
+import org.junit.Assert;
+import org.junit.Test;
+
+public class ProjectHelpersTest {
+
+    @Test
+    public void readUserKey() throws IOException {
+        System.setProperty("user.home",
+                TestUtils.getTestFolder("stats-data").toPath().toString());
+
+        // Read from file
+        String keyString = "user-ab641d2c-test-test-file-223cf1fa628e";
+        String key = ProjectHelpers.getUserKey();
+        assertEquals(keyString, key);
+
+        // Try with non existent
+        File tempDir = File.createTempFile("user.home", "test");
+        tempDir.delete(); // Delete
+        tempDir.mkdir(); // Recreate as directory
+        tempDir.deleteOnExit();
+        File vaadinHome = new File(tempDir, ".vaadin");
+        vaadinHome.mkdir();
+
+        System.setProperty("user.home", tempDir.getAbsolutePath());
+        String newKey = ProjectHelpers.getUserKey();
+        assertNotNull(newKey);
+        assertNotSame(keyString, newKey);
+        File userKeyFile = new File(vaadinHome, "userKey");
+        Assert.assertTrue("userKey should be created automatically",
+                userKeyFile.exists());
+    }
+
+    @Test
+    public void readProKey() {
+        System.setProperty("user.home",
+                TestUtils.getTestFolder("stats-data").toPath().toString());
+
+        // File is used by default
+        String keyStringFile = "test@vaadin.com/pro-536e1234-test-test-file-f7a1ef311234";
+        String keyFile = ProjectHelpers.getProKey();
+        assertEquals(keyStringFile, "test@vaadin.com/" + keyFile);
+
+        // Check system property works
+        String keyStringProp = "test@vaadin.com/pro-536e1234-test-test-prop-f7a1ef311234";
+        System.setProperty("vaadin.proKey", keyStringProp);
+        String keyProp = ProjectHelpers.getProKey();
+        assertEquals(keyStringProp, "test@vaadin.com/" + keyProp);
+    }
+
+}
